created base classes.
Andrea suggested to create base classes so we can share code between
SQLite and MySQL implementations. I kept the CSQLDatabase and 
CSQLPreparedStatement protocols, and created the CSQLDatabase and
CSQLPreparedStatement classes (which respect the protocols).

Marked the messages that must be implemented by the subclasses (or even
the base class) as optional.