using 'int' for consistency
moreover, that column index is going to be used inside a for loop ...
where the caller will probably use something like 'int i'.
I think it's better to keep it a signed integer and not an
'unsigned int' (which is what NSUInteger actually hides).
if unsigned that can lead to annoying warnings or even unexpected
behaviours (but well .. it's difficult to have so many columns to
overflow the signed integer, so warnings and consistency are what
I really care of)