starte cleaning-up things a bit (WIP)
introducing CSQLResultCallback to be used to build the resultset
containers or as facility to access the callback-based api.

Got rid of C callbacks in favour of obj-c selectors

conforming the exec* api to return a boolean value

introduced an internal storage for affectedRows in the base CSDatabase
class. That can be used by drivers to store the number of affected rows
if they are going to release resources as soon as possible so loosing
the possibility to return the number of affected rows later when
requested by the caller

aligned mysql connector to behave like most recent sqlite one
(in terms of api and returned datatypes)